requirements.txt
fastapi>=0.115.13
uvicorn[standard]>=0.34.3
sqlmodel>=0.0.24
sqlalchemy>=2.0.41
psycopg[binary]>=3.2.9
httpx>=0.28.1
openai>=1.90.0
tiktoken>=0.9.0
typer>=0.16.0
prometheus-fastapi-instrumentator>=7.1.0
pgvector>=0.4.1
asyncpg>=0.30.0
pydantic[email]>=2.11.7
pydantic-settings>=2.10.1
tenacity>=9.1.2
asgi-lifespan>=2.1.0
python-dotenv>=1.0.0
alembic>=1.16.2